NormalizeRoute class does not generate URL when the URL has no parameters.
A route with a static URL would match every request for URL generation. To understand why this is the case, read my blog post: [Named routes to the rescue](http://haacked.com/archive/2010/11/21/named-routes-to-the-rescue.aspx)
Changing the the `NormalizeRoute` to not render a URL in this case is technically a breaking change, but it's more likely to fix things than break things. After all, if you have a static URL, why do you need URL generation? :)
Fixes Issue #2